backend-legacy-python/pyproject.toml
[project]
name = "chameleon-vpn-backend"
version = "1.0.0"
description = "Chameleon VPN Backend API"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Redis
    "redis[hiredis]>=5.2",
    # Config
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    # Auth
    "joserfc>=1.0",
    "PyJWT>=2.9",
    "bcrypt>=4.2",
    "httpx>=0.28",
    # Apple StoreKit 2 (official Apple library)
    "app-store-server-library>=3.0",
    # VPN node management
    "paramiko>=3.5",
    # Monitoring
    "psutil>=6.1",
    # Middleware
    "starlette>=0.41",
    "python-multipart>=0.0.18",
    "itsdangerous>=2.2",
    # Data files
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["app"]
omit = ["app/alembic/*", "tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
